Fixed problems with Hydra render delegates during launch of Houdini
There are a couple of issues I noticed with the setting of the Hydra render delegate when Houdini is launched.

First, XUSD_ImagingEngineGL::SetRendererPlugin() is called numerous times with the renderer ping-ponging between the
default renderer (HD_HoudiniRendererPlugin aka Houdini GL), and our custom render delegate (ChmRendererPlugin aka
Chimera). This causes the render delegates (and their associated plugins) to be constructed and destructed multiple
times.

The big problem we were having with this was that during the construction, destruction, and setting of render delegates
(and their associated plugins), a shared context is created which subsequently becomes invalid when the plugin that
created it is destructed. And since the shared context can only be initialized once (see
GlfGLContextRegistry::GetShared()) every time somebody wants to use it, they get an invalid context.

The underlying problems and solutions are twofold:
1. The XUSD_ImagingEngineGL constructor was calling _InitGL() (which creates the shared context) before the
   GlfGLContextRegistry (which is used to create the shared context) had been initialized. The fix for this was
   to move the call to _InitGL() to the end of the constructor.
2. The XUSD_ImagingEngineGL constructor was always setting the renderer plugin to the default (which is
   HD_HoudiniRendererPlugin) and this was later being changed to the desired renderer plugin in
   HUSD_Imaging::setupRenderer(). The fix for this was to pass the desired renderer plugin (name) to the
   XUSD_ImagingEngineGL constructor from HUSD_Imaging::setupRenderer().